feat: socketio server deploy
